Move utils and datatypes to separate files
