+++ ehrapy/tools/_scanpy_tl_api.py
@@ -7,23 +7,11 @@ from anndata import AnnData
 from leidenalg.VertexPartition import MutableVertexPartition
 from scanpy._utils import AnyRandom
 from scipy.sparse import spmatrix
-from scipy.stats import chi2_contingency
 import pandas as pd
 
 from ehrapy.preprocessing._scanpy_pp_api import pca  # noqa: E402,F403,F401
-
-def _merge_arrays(recarray, array, groups_order):
-    """Merge `recarray` obtained from scanpy with manually created numpy `array`"""
-
-    # The easiest way to convert recarray to a normal array is through pandas
-    df = pd.DataFrame(recarray)
-
-    # In case groups have different order
-    converted_recarray = df[groups_order]
-    concatenated_arrays = pd.concat([converted_recarray, pd.DataFrame(array, columns=groups_order)],
-                                    ignore_index=True, axis=0)
-    
-    return concatenated_arrays.to_records(index=False)
+from ehrapy.tools import _datatypes
+from ehrapy.tools import _utils
 
 
 def tsne(
@@ -92,9 +80,6 @@ def tsne(
     )
 
 
-_InitPos = Literal["paga", "spectral", "random"]
-
-
 def umap(
     adata: AnnData,
     min_dist: float = 0.5,
@@ -104,7 +89,7 @@ def umap(
     alpha: float = 1.0,
     gamma: float = 1.0,
     negative_sample_rate: int = 5,
-    init_pos: Union[_InitPos, np.ndarray, None] = "spectral",
+    init_pos: Union[_datatypes._InitPos, np.ndarray, None] = "spectral",
     random_state: AnyRandom = 0,
     a: Optional[float] = None,
     b: Optional[float] = None,
@@ -198,13 +183,9 @@ def umap(
         raise ValueError(f'.uns["{neighbors_key}"] or .uns["neighbors"] were not found. Run `ep.pp.neighbors` first.')
 
 
-_LAYOUTS = ("fr", "drl", "kk", "grid_fr", "lgl", "rt", "rt_circular", "fa")
-_Layout = Literal[_LAYOUTS]  # type: ignore
-
-
 def draw_graph(
     adata: AnnData,
-    layout: _Layout = "fa",
+    layout: _datatypes._Layout = "fa",
     init_pos: Union[str, bool, None] = None,
     root: Optional[int] = None,
     random_state: AnyRandom = 0,
@@ -775,226 +756,6 @@ def ingest(
     )
 
 
-_rank_features_groups_method = Optional[Literal["logreg", "t-test", "wilcoxon", "t-test_overestim_var"]]
-_corr_method = Literal["benjamini-hochberg", "bonferroni"]
-_rank_features_groups_cat_method = Literal["chi-square", "g-test", "freeman-tukey", "mod-log-likelihood", "neyman", "cressie-read"]
-
-
-def _adjust_pvalues(pvals: np.recarray, corr_method: _corr_method):
-    """Perform per group p-values correction with a given `corr_method`
-    
-    Args:
-        pvals: numpy records array with p-values. The resulting p-values are corrected per group (i.e. column)
-        corr_method: p-value correction method
-
-    Returns:
-        Records array of the same format as an input but with corrected p-values
-    """
-    from statsmodels.stats.multitest import multipletests
-
-    method_map = {
-        "benjamini-hochberg": "fdr_bh",
-        "bonferroni": "bonferroni"
-    }
-
-    pvals_adj = np.ones_like(pvals)
-
-    for group in pvals.dtype.names:
-        group_pvals = pvals[group]
-        
-        _, group_pvals_adj, _, _ = multipletests(
-            group_pvals, alpha=0.05, method=method_map[corr_method]
-        )
-        pvals_adj[group] = group_pvals_adj
-
-    return pvals_adj
-
-
-def _sort_features(adata, key_added="rank_features_groups") -> None:
-    """Sort results of :func:`~ehrapy.tl.rank_features_groups` by adjusted p-value
-    
-    Args:
-        adata: Annotated data matrix after running :func:`~ehrapy.tl.rank_features_groups`
-        key_added: The key in `adata.uns` information is saved to.
-
-    Returns:
-        Nothing. The operation is performed in place
-    """
-    if key_added not in adata.uns:
-        return
-
-    pvals_adj = adata.uns[key_added]["pvals_adj"]
-
-    for group in pvals_adj.dtype.names:
-        group_pvals = pvals_adj[group]
-        sorted_indexes = np.argsort(group_pvals)
-        
-        for key in adata.uns[key_added].keys():
-            if key == "params":
-                # This key only stores technical information, nothing to sort here
-                continue
-            
-            # Sort every key (e.g. pvals, names) by adjusted p-value in an increasing order
-            adata.uns[key_added][key][group] = adata.uns[key_added][key][group][sorted_indexes]
-
-def _save_rank_features_result(adata, key_added, names, scores, pvals, pvals_adj=None, logfoldchanges=None, pts=None, groups_order=None) -> None:
-    """Write keys with statistical test results to adata.uns
-    
-    Args:
-        adata: Annotated data matrix after running :func:`~ehrapy.tl.rank_features_groups`
-        key_added: The key in `adata.uns` information is saved to.
-        names: Structured array storing the feature names
-        scores: Array with the statistics
-        logfoldchanges: logarithm of fold changes or other info to store under logfoldchanges key
-        pvals: p-values of a statistical test 
-        pts: Percentages of cells containing features
-        groups_order: order of groups in structured arrays
-
-    Returns:
-        Nothing. The operation is performed in place
-    """
-    fields = (names, scores, pvals, pvals_adj, logfoldchanges, pts)
-    field_names = ("names", "scores", "pvals", "pvals_adj", "logfoldchanges", "pts")
-
-    for values, key in zip(fields, field_names):
-        if values is None or not len(values):
-            continue
-
-        if key not in adata.uns[key_added]:
-            adata.uns[key_added][key] = values
-        else:
-            adata.uns[key_added][key] = _merge_arrays(
-                recarray=adata.uns[key_added][key],
-                array=np.array(values),
-                groups_order=groups_order
-            )
-
-def _get_groups_order(groups_subset, group_names, reference):
-    """Convert `groups` parameter of :func:`~ehrapy.tl.rank_features_groups` to a list of groups
-    
-    Args:
-        groups_subset: Subset of groups, e.g. [`'g1'`, `'g2'`, `'g3'`], to which comparison
-                       shall be restricted, or `'all'` (default), for all groups.
-        group_names: list of all available group names
-        reference: One of the groups of `'rest'`
-
-    Returns:
-        List of groups, subsetted or full
-    """
-    if groups_subset == "all":
-        groups_order = group_names
-    elif isinstance(groups_subset, (str, int)):
-        raise ValueError("Specify a sequence of groups")
-    else:
-        groups_order = list(groups_subset)
-        if isinstance(groups_order[0], int):
-            groups_order = [str(n) for n in groups_order]
-        if reference != "rest" and reference not in groups_order:
-            groups_order += [reference]
-    if reference != "rest" and reference not in group_names:
-        raise ValueError(
-            f"reference = {reference} needs to be one of groupby = {group_names}."
-        )
-    
-    return groups_order
-
-def _evaluate_categorical_features(
-        adata,
-        groupby,
-        group_names,
-        groups: Union[Literal["all"], Iterable[str]] = "all",
-        reference: str = "rest", 
-        categorical_method: _rank_features_groups_cat_method = "g-test", 
-        pts=False
-):
-    """Run statistical test for categorical features
-
-    Args:
-        adata: Annotated data matrix.
-        groupby: The key of the observations grouping to consider.
-        groups: Subset of groups, e.g. [`'g1'`, `'g2'`, `'g3'`], to which comparison
-                shall be restricted, or `'all'` (default), for all groups.
-        reference: If `'rest'`, compare each group to the union of the rest of the group.
-                   If a group identifier, compare with respect to this group.
-        pts: Whether to add 'pts' key to output. Doesn't contain useful information in this case.
-        categorical_method: statistical method to calculate differences between categories
-
-    Returns:
-        *names*: `np.array` 
-                  Structured array to be indexed by group id storing the feature names
-        *scores*: `np.array`
-                  Array to be indexed by group id storing the statistic underlying
-                  the computation of a p-value for each feature for each group.
-        *logfoldchanges*: `np.array`
-                          Always equal to 1 for this function 
-        *pvals*: `np.array`
-                 p-values of a statistical test 
-        *pts*: `np.array`
-                 Always equal to 1 for this function
-    """
-    tests_to_lambdas = {
-        "chi-square": 1,
-        "g-test": 0,
-        "freeman-tukey": -1/2,
-        "mod-log-likelihood": -1,
-        "neyman": -2,
-        "cressie-read": 2/3,
-    }
-
-    categorical_names = []
-    categorical_scores = []
-    categorical_pvals = []
-    categorical_logfoldchanges = []
-    categorical_pts = []
-
-    groups_order = _get_groups_order(groups_subset=groups, group_names=group_names, reference=reference)
-
-    groups_values = adata.obs[groupby].to_numpy()
-
-    for feature in adata.uns["non_numerical_columns"]:
-        if feature == groupby or "ehrapycat_" + feature == groupby or feature == "ehrapycat_" + groupby:
-            continue
-            
-        feature_values = adata[:, feature].X.flatten().toarray()
-
-        pvals = []
-        scores = []
-
-        for group in group_names:
-            if group not in groups_order:
-                continue
-
-            if reference == "rest":
-                reference_mask = (groups_values != group) & np.isin(groups_values, groups_order)
-                contingency_table = pd.crosstab(feature_values, reference_mask)
-            else:
-                obs_to_take = np.isin(groups_values, [group, reference])
-                reference_mask = groups_values[obs_to_take] == reference
-                contingency_table = pd.crosstab(feature_values[obs_to_take], reference_mask)
-
-            score, p_value, _, _ = chi2_contingency(
-                contingency_table.values, lambda_=tests_to_lambdas[categorical_method])
-            scores.append(score)
-            pvals.append(p_value)
-        
-        categorical_names.append([feature] * len(group_names))
-        categorical_scores.append(scores)
-        categorical_pvals.append(pvals)
-        # It is not clear, how to interpret logFC or percentages for categorical data
-        # For now, leave some values so that plotting and sorting methods work
-        categorical_logfoldchanges.append(np.ones(len(group_names)))
-        if pts:
-            categorical_pts.append(np.ones(len(group_names)))
-
-    return (
-        np.array(categorical_names), 
-        np.array(categorical_scores),
-        np.array(categorical_pvals),
-        np.array(categorical_logfoldchanges),
-        np.array(categorical_pts)
-    )
-
-
 def rank_features_groups(
     adata: AnnData,
     groupby: str,
@@ -1005,9 +766,9 @@ def rank_features_groups(
     pts: bool = False,
     key_added: Optional[str] = "rank_features_groups",
     copy: bool = False,
-    method: _rank_features_groups_method = None,
-    categorical_method: _rank_features_groups_cat_method = "g-test",
-    corr_method: _corr_method = "benjamini-hochberg",
+    method: _datatypes._rank_features_groups_method = None,
+    categorical_method: _datatypes._rank_features_groups_cat_method = "g-test",
+    corr_method: _datatypes._corr_method = "benjamini-hochberg",
     tie_correct: bool = False,
     layer: Optional[str] = None,
     **kwds,
@@ -1114,7 +875,7 @@ def rank_features_groups(
         )
 
         # Update adata.uns with numerical result
-        _save_rank_features_result(
+        _utils._save_rank_features_result(
             adata,
             key_added,
             names=numerical_adata.uns[key_added]["names"],
@@ -1127,7 +888,7 @@ def rank_features_groups(
         )
     
     if adata.uns["non_numerical_columns"]:
-        categorical_names, categorical_scores, categorical_pvals, categorical_logfoldchanges, categorical_pts = _evaluate_categorical_features(
+        categorical_names, categorical_scores, categorical_pvals, categorical_logfoldchanges, categorical_pts = _utils._evaluate_categorical_features(
             adata=adata,
             groupby=groupby,
             group_names=group_names,
@@ -1136,7 +897,7 @@ def rank_features_groups(
             categorical_method=categorical_method
         )
 
-        _save_rank_features_result(
+        _utils._save_rank_features_result(
             adata,
             key_added,
             names=categorical_names,
@@ -1150,13 +911,13 @@ def rank_features_groups(
             
     # Adjust p values  
     if "pvals" in adata.uns[key_added]:
-        adata.uns[key_added]["pvals_adj"] = _adjust_pvalues(adata.uns[key_added]["pvals"], corr_method=corr_method)
+        adata.uns[key_added]["pvals_adj"] = _utils._adjust_pvalues(adata.uns[key_added]["pvals"], corr_method=corr_method)
         
     # For some reason, pts should be a DataFrame
     if "pts" in adata.uns[key_added]:
         adata.uns[key_added]["pts"] = pd.DataFrame(adata.uns[key_added]["pts"])
 
-    _sort_features(adata, key_added)
+    _utils._sort_features(adata, key_added)
 
     return adata if copy else None
 
@@ -1210,15 +971,12 @@ def filter_rank_features_groups(
     )
 
 
-_marker_feature_overlap_methods = Literal["overlap_count", "overlap_coef", "jaccard"]
-
-
 def marker_feature_overlap(
     adata: AnnData,
     reference_markers: Union[Dict[str, set], Dict[str, list]],
     *,
     key: str = "rank_features_groups",
-    method: _marker_feature_overlap_methods = "overlap_count",
+    method: _datatypes._marker_feature_overlap_methods = "overlap_count",
     normalize: Optional[Literal["reference", "data"]] = None,
     top_n_markers: Optional[int] = None,
     adj_pval_threshold: Optional[float] = None,
