+++ tests/tools/test_features_ranking.py
@@ -2,6 +2,7 @@ import numpy as np
 import pandas as pd
 
 import ehrapy as ep
+from ehrapy.tools import _utils
 
 
 class TestHelperFunctions:
@@ -10,12 +11,12 @@ class TestHelperFunctions:
         expected_result_bh = pd.DataFrame({"group1": (0.04, 0.04, 0.04, 1.00), "group2": (0.08, 0.08, 0.08, 0.99)}).to_records()
         expected_result_bf = pd.DataFrame({"group1": (0.04, 0.08, 0.12, 1.00), "group2": (0.16, 0.20, 0.24, 1.00)}).to_records()
 
-        result_bh = ep.tl._adjust_pvalues(pvals, method="benjamini-hochberg")
+        result_bh = _utils._adjust_pvalues(pvals, method="benjamini-hochberg")
         assert pvals["group1"] <= result_bh["group1"]
         assert pvals["group2"] <= result_bh["group2"]
         assert np.allclose(result_bh, expected_result_bh)
         
-        result_bf = ep.tl._adjust_pvalues(pvals, method="bonferroni")
+        result_bf = _utils._adjust_pvalues(pvals, method="bonferroni")
         assert pvals["group1"] <= result_bf["group1"]
         assert pvals["group2"] <= result_bf["group2"]
         assert np.allclose(result_bf, expected_result_bf)
@@ -37,7 +38,7 @@ class TestHelperFunctions:
         # Doesn't really matter that they are the same here but order should be preserved
         adata.uns["rank_features_groups"]["pvals_adj"] = adata.uns["rank_features_groups"]["pvals"]
 
-        ep.tl._sort_features(adata)
+        _utils._sort_features(adata)
 
         # Check that every feature is sorted
         # Actually, some of them would be sorted in the opposite direction (e.g. scores) for real data
